Fixed error when using within route
Issue:
When the notifier service gets initialized, in any component which is not the app component, the root component is already here - means, it is already bootstraped. Therefore the bootstrap event handler for getting the root component never fires, and the notifier container never gets build. This breaks all functionality.

Solution:
The notifier service constructor now differenciates between the two cases (whether the root component is already here OR not built yet). As a consequence, the library should now work fine when used within routes.